requirements.txt
nltk==3.8.1
numpy==1.26.2
pandas==2.1.4
PyYAML==6.0.1
scikit_learn==1.3.2
torch==2.1.1
tqdm==4.66.1
transformers==4.36.0
